cleaned up word types
